--- src/ENode.EQueue/PublishableExceptions/PublishableExceptionMessage.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-namespace ENode.EQueue
-{
-    [Serializable]
-    public class PublishableExceptionMessage
-    {
-        public string UniqueId { get; set; }
-        public DateTime Timestamp { get; set; }
-        public IDictionary<string, string> Items { get; set; }
-        public IDictionary<string, string> SerializableInfo { get; set; }
-    }
-}
