+++ src/ENode.Tests/Providers/DomainEventTopicProvider.cs
@@ -1,0 +1,13 @@
+﻿using ENode.EQueue;
+using ENode.Eventing;
+
+namespace ENode.Tests
+{
+    public class DomainEventTopicProvider : AbstractTopicProvider<IDomainEvent>
+    {
+        public override string GetTopic(IDomainEvent source)
+        {
+            return "EventTopic";
+        }
+    }
+}
