+++ src/ENode.Tests/Providers/DomainExceptionTopicProvider.cs
@@ -1,0 +1,13 @@
+﻿using ENode.Domain;
+using ENode.EQueue;
+
+namespace ENode.Tests
+{
+    public class DomainExceptionTopicProvider : AbstractTopicProvider<IDomainException>
+    {
+        public override string GetTopic(IDomainException source)
+        {
+            return "DomainExceptionTopic";
+        }
+    }
+}
