+++ src/Samples/BankTransferSample/TopicProviders/ExceptionTopicProvider.cs
@@ -1,11 +1,11 @@
-﻿using ENode.EQueue;
-using ENode.Infrastructure;
+﻿using ENode.Domain;
+using ENode.EQueue;
 
 namespace BankTransferSample.Providers
 {
-    public class ExceptionTopicProvider : AbstractTopicProvider<IPublishableException>
+    public class ExceptionTopicProvider : AbstractTopicProvider<IDomainException>
     {
-        public override string GetTopic(IPublishableException source)
+        public override string GetTopic(IDomainException source)
         {
             return Constants.ExceptionTopic;
         }
